Create component for the change display name form
